requirements.txt
pytesseract~=0.3.13
pillow~=11.0.0